projects/lecture_video_generator/requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
python-docx>=1.0.0
svgwrite>=1.4.0
CairoSVG>=2.7.0
Jinja2>=3.1.0
google-genai
